+++ tests/internal/notices/ad-hoc/paper-returns.spec.js
@@ -1,4 +1,4 @@
-import scenarioData from '../../../support/scenarios/licence-with-previous-return-log.js'
+import scenarioData from '../../../support/scenarios/licence-with-previous-return-log.scenario.js'
 import { test, expect } from '../../../support/fixtures.js'
 
 test.describe('Ad-hoc Paper returns journey (internal)', () => {
