+++ tests/support/scenarios/licence-with-open-return-log-bad-primary-user.scenario.js
@@ -1,11 +1,9 @@
-import company from '../data/company.js'
-import licence from '../data/licence.js'
-import point from '../data/point.js'
-import primaryUser from '../data/primary-user.js'
-import returnLog from '../data/return-log.js'
-import returnRequirement from '../data/return-requirement.js'
-import returnVersion from '../data/return-version.js'
+import registeredLicenceScenario from './registered-licence.scenario.js'
+import returnLogData from '../data/return-log.data.js'
+import returnRequirementData from '../data/return-requirement.data.js'
+import returnVersionData from '../data/return-version.data.js'
 import { compareDates, previousPeriod, today } from '../helpers/date.helpers.js'
+import { mergeByKey } from '../helpers/scenario.helpers.js'
 
 export const title = 'Open return log with bad primary user'
 export const description =
@@ -23,13 +21,11 @@ export default function (calculatedDates) {
     returnPeriod = previousPeriod(firstReturnPeriod)
   }
 
-  const licenceRef = 'AT/TE/ST/01/01'
-
-  const companyData = company()
+  const registeredLicence = registeredLicenceScenario('AT/TE/ST/01/01', 'iwill-fail@e')
 
   const {
     addresses: [address]
-  } = companyData
+  } = registeredLicence
 
   // We'll only set the due date on the OPEN return log if the alternate notification is successful. The Notify
   // service will reject the request if the address is not real, even though we're using our Notify test API key.
@@ -42,25 +38,14 @@ export default function (calculatedDates) {
   address.address6 = null
   address.postcode = 'BS1 5AH'
 
-  const primaryUserData = primaryUser('iwill-fail@e', companyData)
-  const licenceData = licence(licenceRef, companyData, primaryUserData)
-  const pointData = point()
-  const returnVersionData = returnVersion(licenceData)
-  const returnRequirementData = returnRequirement(returnVersionData, pointData)
-  const returnLogData = returnLog(licenceData, returnRequirementData, pointData, {
+  const returnVersion = returnVersionData(registeredLicence)
+  const returnRequirement = returnRequirementData(returnVersion, registeredLicence)
+  const returnLog = returnLogData(registeredLicence, returnRequirement, {
     startDate: returnPeriod.startDate,
     endDate: returnPeriod.endDate,
     dueDate: null,
     quarterly: returnPeriod.quarterly
   })
 
-  return {
-    ...companyData,
-    ...primaryUserData,
-    ...licenceData,
-    ...pointData,
-    ...returnVersionData,
-    ...returnRequirementData,
-    ...returnLogData
-  }
+  return mergeByKey(registeredLicence, returnVersion, returnRequirement, returnLog)
 }
