+++ tests/support/scenarios/registered-licence.scenario.js
@@ -1,19 +1,18 @@
-import companyData from '../data/company.data.js'
-import licenceData from '../data/licence.data.js'
-import pointData from '../data/point.data.js'
 import primaryUserData from '../data/primary-user.data.js'
+import unregisteredLicenceScenario from './unregistered-licence.scenario.js'
 import { mergeByKey } from '../helpers/scenario.helpers.js'
 
 export const title = 'Licence (registered)'
 export const description = 'A licence that has been registered (primary user), licence holder and a company'
 
-export default function () {
-  const licenceRef = 'AT/TE/ST/01/01'
+export default function (licenceRef = 'AT/TE/ST/01/01', email = 'external@example.com') {
+  const unregisteredLicence = unregisteredLicenceScenario(licenceRef)
 
-  const company = companyData()
-  const point = pointData()
+  const primaryUser = primaryUserData(email, unregisteredLicence)
 
-  const primaryUser = primaryUserData('external@example.com', company)
+  // Linking a primary user's company entity to the licence's 'licenceDocumentHeaders' is the only way we can link a
+  // registered licence to a licence holder.
+  unregisteredLicence.licenceDocumentHeaders[0].companyEntityId = primaryUser.licenceEntityRoles[0].companyEntityId
 
-  return mergeByKey(company, primaryUser, point, licenceData(licenceRef, company, point, primaryUser))
+  return mergeByKey(unregisteredLicence, primaryUser)
 }
